* fixed compilation of ufo2map

+++ src/tools/ufo2map/writebsp.c
@@ -72,8 +72,8 @@ static void EmitLeaf (const node_t *node)
 	leaf_p->area = node->area;
 
 	/* write bounding box info */
-	VectorCopy((short)node->mins, leaf_p->mins);
-	VectorCopy((short)node->maxs, leaf_p->maxs);
+	VectorCopy(node->mins, leaf_p->mins);
+	VectorCopy(node->maxs, leaf_p->maxs);
 
 	/* write the leafbrushes */
 	leaf_p->firstleafbrush = curTile->numleafbrushes;
@@ -158,8 +158,8 @@ static int EmitDrawNode_r (node_t *node)
 	Verb_Printf(VERB_DUMP, "EmitDrawNode_r: creating bsp node %i\n", curTile->numnodes);
 	curTile->numnodes++;
 
-	VectorCopy((short)node->mins, n->mins);
-	VectorCopy((short)node->maxs, n->maxs);
+	VectorCopy(node->mins, n->mins);
+	VectorCopy(node->maxs, n->maxs);
 
 	if (node->planenum & 1)
 		Sys_Error("EmitDrawNode_r: odd planenum: %i", node->planenum);
